% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_ebsco.R
\name{fetch_ebsco}
\alias{fetch_ebsco}
\title{Pubmed fetch one page (up to 500 refs) from search}
\usage{
fetch_ebsco(pagenumber, historyinfo)
}
\arguments{
\item{pagenumber}{number of page to retrieve}

\item{historyinfo}{web environment info returned from search_pm()}
}
\value{
tibble with info fields on up to 500 articles
}
\description{
Pubmed fetch one page (up to 500 refs) from search
}
