% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_search.R
\name{fetch_pm}
\alias{fetch_pm}
\title{Pubmed fetch one page (up to 500 refs) from search}
\usage{
fetch_pm(pagenumber, historyinfo)
}
\arguments{
\item{pagenumber}{number of page to retrieve}

\item{historyinfo}{web environment info returned from search_pm()}
}
\value{
tibble with info fields on up to 500 articles
}
\description{
Pubmed fetch one page (up to 500 refs) from search
}
