% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{format_month_abb_year}
\alias{format_month_abb_year}
\title{Format dates with month abbreviation and year}
\usage{
format_month_abb_year(s)
}
\arguments{
\item{s}{string in the format Mar2020}
}
\value{
a string
}
\description{
Format dates from Mar2020 to 2020-03-01
}
