% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{format_month_abb_year_v}
\alias{format_month_abb_year_v}
\title{Format dates with month abbreviation and year vectorized}
\usage{
format_month_abb_year_v(s)
}
\arguments{
\item{s}{a vector of strings}
}
\value{
a string
}
\description{
Format vector of dates from Mar2020 to 2020-03-01
}
