% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_ebsco.R
\name{gen_url_ebsco}
\alias{gen_url_ebsco}
\title{Generates URL with query string for Ebsco API}
\usage{
gen_url_ebsco(searchterm, datefrom = Sys.Date() - 365, dateto = Sys.Date() - 1)
}
\arguments{
\item{searchterm}{text of the query}

\item{datefrom}{from date appeared online (default = 1 year ago from today), format YYYY/MM/DD}

\item{dateto}{to date appeared online (default = today), format YYYY/MM/DD}
}
\value{
string, a URL to search pubmed for a particular term between two given dates
}
\description{
Generates URL with query string for Ebsco API
}
