% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searches.R
\name{gen_url_pm}
\alias{gen_url_pm}
\title{Generates URL for pubmed API}
\usage{
gen_url_pm(searchterm, datefrom = Sys.Date() - 365, dateto = Sys.Date())
}
\arguments{
\item{searchterm}{text of the query}

\item{datefrom}{from date appeared online (default = 1 year ago today), format YYYY/MM/DD}

\item{dateto}{to date appeared online (default = today), format YYYY/MM/DD}
}
\value{
string, a URL to search pubmed for a particular term between two given dates
}
\description{
Generates URL for pubmed API
}
