% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_scopus.R
\name{gen_url_scopus}
\alias{gen_url_scopus}
\title{Scopus generate URL}
\usage{
gen_url_scopus(
  searchterm,
  datefrom = Sys.Date() - 365,
  dateto = Sys.Date() - 1,
  cursor = "*"
)
}
\arguments{
\item{searchterm}{string with search term}

\item{datefrom}{search articles published from (default one year ago)}

\item{dateto}{search articles published until (default today)}

\item{cursor}{code for cursor (defaults to asterisk for first page)}
}
\value{
a string with the URL to hit
}
\description{
Scopus generate URL
}
