% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_ebsco.R
\name{get_ebsco}
\alias{get_ebsco}
\title{ebsco get page of results}
\usage{
get_ebsco(searchterm, datefrom = Sys.Date() - 365, dateto = Sys.Date() - 1)
}
\arguments{
\item{searchterm}{text of the query}

\item{datefrom}{from date appeared online (default = 1 year ago from today), format YYYY/MM/DD}

\item{dateto}{to date appeared online (default = today), format YYYY/MM/DD}
}
\value{
a tibble of all results
}
\description{
ebsco get page of results
}
