% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_ebsco.R
\name{get_number_of_hits_ebsco}
\alias{get_number_of_hits_ebsco}
\title{Get the number of results from querying Ebsco (fsr) db}
\usage{
get_number_of_hits_ebsco(url)
}
\arguments{
\item{url}{a string, url generated by gen_url_ebsco}
}
\value{
a numerical value of the total number of results found
}
\description{
To get the number of hits and get a faster response time, limit returned
results to 1.
}
