% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_search.R
\name{get_pm}
\alias{get_pm}
\title{Pubmed all steps}
\usage{
get_pm(searchterm, datefrom = Sys.Date() - 365, dateto = Sys.Date() - 1)
}
\arguments{
\item{searchterm}{text of the query}

\item{datefrom}{from date appeared online (default = 1 year ago today), format YYYY/MM/DD}

\item{dateto}{to date appeared online (default = today), format YYYY/MM/DD}
}
\value{
a tibble of results
}
\description{
Pubmed all steps
}
