% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_springer.R
\name{get_results_springer}
\alias{get_results_springer}
\title{Springer fetch 1 page (up to 100 per page)}
\usage{
get_results_springer(page, searchurl)
}
\arguments{
\item{page}{page to fetch}

\item{searchurl}{URL with the search query}
}
\description{
Springer fetch 1 page (up to 100 per page)
}
