% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_search.R
\name{get_scopus}
\alias{get_scopus}
\title{Scopus all steps}
\usage{
get_scopus(
  searchterm,
  dateto = Sys.Date() - 1,
  datefrom = Sys.Date() - 365,
  cursor = "*"
)
}
\arguments{
\item{searchterm}{string with search term}

\item{dateto}{search articles published until (default today)}

\item{datefrom}{search articles published from (default one year ago)}

\item{cursor}{code for cursor (defaults to asterisk for first page)}
}
\value{
a tibble with all results
}
\description{
Scopus all steps
}
