% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_scopus.R
\name{get_scopus_result}
\alias{get_scopus_result}
\title{Scopus get page of results}
\usage{
get_scopus_result(url)
}
\arguments{
\item{url}{URL to hit}
}
\value{
a list with a tibble of results, a URL for the next page, and the total number of pages
}
\description{
Scopus get page of results
}
