% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_search.R
\name{getab}
\alias{getab}
\title{Scopus missing abstracts}
\usage{
getab(doi)
}
\arguments{
\item{doi}{string of DOI of article abstract to retrieve}
}
\value{
a tibble with all results
}
\description{
Fetches forbidden scopus abstracts from sciencedirect
}
