% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_search.R
\name{gettotal}
\alias{gettotal}
\title{Search counter}
\usage{
gettotal(searchterm, dateto = Sys.Date(), datefrom = Sys.Date() - 365, across)
}
\arguments{
\item{searchterm}{string of search term}

\item{dateto}{search articles published until (default today)}

\item{datefrom}{search articles published from (default one year ago)}
}
\description{
gets a total search count across all sources
}
