% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_search.R
\name{search_pm}
\alias{search_pm}
\title{Pubmed search}
\usage{
search_pm(searchurl)
}
\arguments{
\item{searchurl}{search URL generated by gen_url_pm()}
}
\value{
A list with the total hits and history parameters for returning article info
}
\description{
Pubmed search
}
