% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_ebsco.R
\name{fetch_ebsco}
\alias{fetch_ebsco}
\title{Ebsco fetch one page (up to 200 refs) from search}
\usage{
fetch_ebsco(url, startrec)
}
\arguments{
\item{url}{a string, url generated by gen_url_ebsco}

\item{startrec}{number starting record number for the result set returned from a search}
}
\value{
tibble with info fields on up to 500 articles
}
\description{
Ebsco fetch one page (up to 200 refs) from search
}
