% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_ebsco.R
\name{get_ebsco}
\alias{get_ebsco}
\title{ebsco get page of results}
\usage{
get_ebsco(searchterm, datefrom = Sys.Date() - 365, dateto = Sys.Date() - 1)
}
\arguments{
\item{url}{URL to hit}
}
\value{
a list with a tibble of results, a URL for the next page, and the total number of pages
}
\description{
ebsco get page of results
}
