% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searches.R
\name{get_results_springer}
\alias{get_results_springer}
\title{Springer fetch 1 page (up to 100 per page)}
\usage{
get_results_springer(page, searchurl)
}
\arguments{
\item{searchurl}{URL with the search query}

\item{searchterm}{original search term (for final filtering)}
}
\description{
Springer fetch 1 page (up to 100 per page)
}
