% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettotal.R
\name{gettotal}
\alias{gettotal}
\title{Search counter}
\usage{
gettotal(
  searchterm,
  dateto = Sys.Date() - 1,
  datefrom = Sys.Date() - 365,
  across
)
}
\arguments{
\item{searchterm}{string of search term}

\item{dateto}{search articles published until (default today)}

\item{datefrom}{search articles published from (default one year ago)}

\item{across}{list of APIs sources to be used}
}
\description{
gets a total search count across all sources
}
