% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_search.R
\name{xml2tib}
\alias{xml2tib}
\title{Pick out nodes from xml and turn it into a tibble}
\usage{
xml2tib(xmlnode, nodenames)
}
\arguments{
\item{xmlnode}{an xml node containing nodes}

\item{nodenames}{the names of the nodes you wish to extract (string with nodes separated by commas)}
}
\value{
A tibble with node names as col names and node text as values
}
\description{
Pick out nodes from xml and turn it into a tibble
}
